requirements.txt
discord.py
requests
selenium
sqlite3
datetime
os
threading